+++ packages/swapkit/helpers/src/helpers/asset.ts
@@ -84,11 +84,11 @@ export const isGasAsset = ({ chain, symbol }: { chain: Chain; symbol: string }) 
     case Chain.Dogecoin:
     case Chain.Binance:
     case Chain.Ethereum:
-    case Chain.Base:
     case Chain.Avalanche:
       return symbol === chain;
     case Chain.Arbitrum:
     case Chain.Optimism:
+    case Chain.Base:
       return 'ETH' === symbol;
     case Chain.Maya:
       return symbol === 'CACAO';
@@ -124,6 +124,8 @@ export const getCommonAssetInfo = (
       return { identifier: 'MAYA.CACAO', decimal: BaseDecimal.MAYA };
     case 'MAYA.MAYA':
       return { identifier: 'MAYA.MAYA', decimal: 4 };
+    case Chain.Base:
+      return { identifier: 'BASE.ETH', decimal: 18 };
     case Chain.Ripple:
     case Chain.Kujira:
     case Chain.Arbitrum:
@@ -137,7 +139,6 @@ export const getCommonAssetInfo = (
     case Chain.Avalanche:
     case Chain.Polygon:
     case Chain.Bitcoin:
-    case Chain.Base:
     case Chain.Ethereum:
       return { identifier: `${assetString}.${assetString}`, decimal: BaseDecimal[assetString] };
   }