+++ packages/swapkit/tokens/src/tokenLists/Native.ts
@@ -5,6 +5,7 @@ export const list = {
   keywords: ['Native'],
   tokens: [
     { chain: 'AVAX', identifier: 'AVAX.AVAX', decimals: 18 },
+    { chain: 'BASE', identifier: 'BASE.ETH', decimals: 18 },
     { chain: 'BCH', identifier: 'BCH.BCH', decimals: 8 },
     { chain: 'BNB', identifier: 'BNB.BNB', decimals: 8 },
     { chain: 'BSC', identifier: 'BSC.BNB', decimals: 18 },