+++ e2e/swaps/e2e-swap-eth-xrp/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @coinmasters/e2e-swap-osmosis-cosmos
 
+## 1.4.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.4.0
+  - @coinmasters/wallet-keystore@8.0.0
+  - @coinmasters/wallet-keepkey@10.0.0
+  - @coinmasters/types@4.4.0
+  - @coinmasters/core@8.0.0
+
 ## 1.3.0
 
 ### Minor Changes