+++ e2e/wallets/intergration-db/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @coinmasters/integration-coins
 
+## 1.4.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-db@4.4.0
+  - @coinmasters/core@8.0.0
+
 ## 1.3.0
 
 ### Minor Changes