+++ internal/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @internal/config
 
+## 2.4.0
+
+### Minor Changes
+
+- bump
+
 ## 2.3.0
 
 ### Minor Changes