+++ packages/swapkit/core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @coinmasters/core
 
+## 8.0.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.0
+  - @coinmasters/helpers@8.0.0
+  - @coinmasters/tokens@3.4.0
+  - @coinmasters/toolbox-utxo@8.0.0
+  - @coinmasters/types@4.4.0
+  - @coinmasters/toolbox-evm@8.0.0
+  - @coinmasters/api@3.4.0
+
 ## 7.0.0
 
 ### Minor Changes