+++ packages/swapkit/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/tokens
 
+## 3.4.0
+
+### Minor Changes
+
+- bump
+
 ## 3.3.0
 
 ### Minor Changes