+++ packages/swapkit/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/types
 
+## 4.4.0
+
+### Minor Changes
+
+- bump
+
 ## 4.3.0
 
 ### Minor Changes