+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @swapkit/toolbox-utxo
 
+## 8.0.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/helpers@8.0.0
+  - @coinmasters/types@4.4.0
+
 ## 7.0.0
 
 ### Minor Changes