+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @coinmasters/wallet-evm-extensions
 
+## 8.0.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.0
+  - @coinmasters/toolbox-evm@8.0.0
+
 ## 7.0.0
 
 ### Minor Changes