+++ packages/wallets/wc/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @coinmasters/wallet-wc
 
+## 8.0.0
+
+### Minor Changes
+
+- bump
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.0
+  - @coinmasters/types@4.4.0
+  - @coinmasters/toolbox-evm@8.0.0
+
 ## 7.0.0
 
 ### Minor Changes